+++ src/mantle/testing/TestCase.php
@@ -26,6 +26,7 @@ use Mantle\Testing\Concerns\Interacts_With_Container;
 use Mantle\Testing\Concerns\Interacts_With_Cron;
 use Mantle\Testing\Concerns\Interacts_With_Hooks;
 use Mantle\Testing\Concerns\Interacts_With_Mail;
+use Mantle\Testing\Concerns\Interacts_With_PHPUnit;
 use Mantle\Testing\Concerns\Interacts_With_Requests;
 use Mantle\Testing\Concerns\Makes_Http_Requests;
 use Mantle\Testing\Concerns\Network_Admin_Screen;
@@ -60,6 +61,7 @@ abstract class TestCase extends BaseTestCase {
 	use Interacts_With_Cron;
 	use Interacts_With_Hooks;
 	use Interacts_With_Mail;
+	use Interacts_With_PHPUnit;
 	use Interacts_With_Requests;
 	use Makes_Http_Requests;
 	use MatchesSnapshots;