+++ CHANGELOG.md
@@ -11,6 +11,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added a `list()`, `ordered_list()`, `reusable()`, and `button()` method to the
   block factory to generate the corresponding blocks.
+- Add support for PHPUnit 12. Note: PHPUnit 12 drops all support for docblock
+  annotations. To easily upgrade your project to use attributes, try
+  [Rector](https://getrector.com/documentation).
+
+### Changed
+
+- When using the `Hookable` trait, any `action__{method}`/`on_{method}` will be
+  ignored when the method uses a `Action`/`Filter` attribute. To allow for
+  legacy behavior, the
+  `Mantle\Support\Attributes\Hookable\Allow_Legacy_Duplicate_Registration`
+  attribute can be used on the class to allow for duplicate registration of the
+  methods.
 
 ## v1.5.8